Merged PR 15052: Catalog current loggers in use
This is an initial attempt at documenting the loggers in use for logs and metrics. Will add to this as more research is done. Feel free to add to it. The purpose is to document what is out there and the long term goal is list of the requirements and arrive a set of common solutions.